restructure methods and exports
