+++ std/jwt/base64/base64url.ts
@@ -2,9 +2,8 @@ import {
   convertBase64ToUint8Array,
   convertUint8ArrayToBase64,
 } from "./base64.ts";
-import {
-  addPaddingToBase64url,
-} from "../deps.ts";
+import { addPaddingToBase64url } from "../../encoding/base64url.ts";
+
 
 function convertBase64urlToBase64(base64url: string): string {
   return addPaddingToBase64url(base64url).replace(/\-/g, "+").replace(
