--- std/jwt/deps.ts
@@ -1,7 +1,0 @@
-export {
-  decodeString as convertHexToUint8Array,
-  encodeToString as convertUint8ArrayToHex,
-} from "../encoding/hex.ts";
-export { HmacSha256 } from "../hash/sha256.ts";
-export { HmacSha512 } from "../hash/sha512.ts";
-export { addPaddingToBase64url } from "../encoding/base64url.ts";
