+++ std/jwt/mod.ts
@@ -1,13 +1,3 @@
-export { create, encrypt, makeSignature, setExpiration } from "./create.ts";
-
-export type { Payload } from "./create.ts";
-
-export {
-  checkHeaderCrit,
-  parseAndDecode,
-  validate,
-  validateObject,
-  verifySignature,
-} from "./validate.ts";
-
-export type { Handlers, Validation } from "./validate.ts";
+export { create } from "./create.ts";
+export { validate } from "./validate.ts";
+export { setExpiration } from "./_util.ts"
